requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
bcrypt==4.1.2
python-dotenv==1.0.0
RPi.GPIO==0.7.1
